requirements.txt
flask>=2.0.0
flask-socketio>=5.0.0
flask-jwt-extended>=4.0.0
flask-cors>=4.0.0
pymongo>=4.0.0
scikit-learn>=1.0.0
ortools>=9.0.0
requests>=2.25.0
python-dotenv>=0.19.0
groq>=0.4.0
numpy>=1.21.0
Werkzeug>=2.0.0